Move contacts tab to a separate file in the browser directory.
